app/requirements.txt
Flask
Flask-HTTPAuth
Flask-Bootstrap
pyepics
